requirements.txt
Flask==2.1.2
requests==2.28.0
SQLAlchemy==1.4.37
python-dateutil~=2.8.2